Cart Page done

+++ AndroidApp/Flutter/qord_app/lib/api/json_handle/menu_item.dart
@@ -2,7 +2,7 @@ class MenuItem {
   String itemName;
   int itemCalories;
   int itemWeight;
-  int itemPrice;
+  double itemPrice;
   List<String> ingredients;
   List<String> allergens;
